Add confirmation for deletion

+++ ConcurSolutionz/ConcurSolutionz/Views/EntryPage.xaml.cs
@@ -367,8 +367,13 @@ public partial class EntryPage : ContentPage
 
     private async void DeleteEntry_Clicked(object sender, EventArgs e)
     {
-        Database.Database.DeleteDirectoryByFilePath(filePath);
-        await Shell.Current.GoToAsync(nameof(MainPage));
+        bool answer = await DisplayAlert("Confirm Deletion", $"Are you sure you want to delete Entry {Path.GetFileNameWithoutExtension(entry.FileName)}?", "Yes", "No");
+
+        if (answer)
+        {
+            Database.Database.DeleteDirectoryByFilePath(filePath);
+            await Shell.Current.GoToAsync(nameof(MainPage));
+        }
     }
 
 
@@ -395,21 +400,25 @@ public partial class EntryPage : ContentPage
 
     }
 
-    private void DeleteRecord_Clicked(object sender, EventArgs e)
+    private async void DeleteRecord_Clicked(object sender, EventArgs e)
     {
         Models.Receipt selectedReceipt = recordCollection.SelectedItem as Models.Receipt;
         Database.Record rec = receipts.FirstOrDefault(r => r.RecordID == selectedReceipt.recordID - 1);
         Database.Receipt receipt = receipts.FirstOrDefault(r => r.RecordID == selectedReceipt.recordID - 1);
         if (selectedReceipt != null)
         {
-            // Remove receipt from collection
-            ReceiptView.Remove(selectedReceipt);
-            receipts.Remove(receipt);
-            entry.DelRecord(rec);
+            bool answer = await DisplayAlert("Confirm Deletion", $"Are you sure you want to delete Record {receipt.RecordID + 1}?", "Yes", "No");
 
-            // Update remaining budget and remove from entry
-            PopulateEntry();
+            if (answer)
+            {
+                // Remove receipt from collection
+                ReceiptView.Remove(selectedReceipt);
+                receipts.Remove(receipt);
+                entry.DelRecord(rec);
 
+                // Update remaining budget and remove from entry
+                PopulateEntry();
+            }
 
         }
     }
